move beanDefinitions up to BeanConfiguration.
they are not part of the declarative nature of DeclarativeBeanConfiguration.
