+++ android-beans/src/main/kotlin/rocks/frieler/android/beans/BeanConfiguration.kt
@@ -1,5 +1,7 @@
 package rocks.frieler.android.beans
 
+import kotlin.reflect.KClass
+
 /**
  * Abstract super-class to define beans for the context of an application.
  *
@@ -11,10 +13,27 @@ package rocks.frieler.android.beans
  * @author Christopher Frieler
  */
 abstract class BeanConfiguration {
+	private val beanDefinitions: MutableList<BeanDefinition<*>> = ArrayList()
+
 	/**
 	 * Returns [BeanDefinition]s for the beans defined by this [BeanConfiguration].
 	 *
 	 * @return a list of [BeanDefinition]s
 	 */
-	abstract fun getBeanDefinitions() : List<BeanDefinition<*>>
+	open fun getBeanDefinitions() : List<BeanDefinition<*>> = beanDefinitions
+
+	/**
+	 * Creates and adds a [BeanDefinition] from the given parts.
+	 *
+	 * @param name the (optional) name of the bean to define
+	 * @param type the type of bean to define
+	 * @param definition the function to create the actual bean instance
+	 * @param T the type of bean to define
+	 * @return a [BeanDefinition] from the given parts
+	 */
+	fun <T : Any> addBeanDefinition(name: String?, type: KClass<T>, definition: (BeansProvider) -> T): BeanDefinition<T> {
+		val beanDefinition = BeanDefinition(name, type, definition)
+		beanDefinitions.add(beanDefinition)
+		return beanDefinition
+	}
 }
